requirements.txt
requests
sentence-transformers
PyMuPDF
numpy
faiss-cpu
openai
streamlit